feat: extract timestamp predicates from generic expressions
It turns out that most of our chunk filtering looks at the timestamp
ranges within `Predicate`. While this if filled for InfluxRPC queries,
we don't have this information for SQL queries.

I see two ways to fix that:

- **Advanced Chunk Filter:** Teach our chunks to evaluate generic
  datafusion expressions. This would have the advantage that we don't
  rely on IOx-specific magic too much. On the other hand, I am unsure
  how much complexity that would require.
- **Lower Generic Expressions for IOx:** Try to scan datfusion
  expressions for simple timestamp filters. Perform this conversion
  early (i.e. when building `Predicate`) so that all chunks can use
  simple timestamp lookups.

Here I try the 2nd approach. Tests are missing for now, since this is
rather an RFC.