feat: event emission subsystem
Add a framework to simplify event emission. This is required for #5464.
Mostly it allows us to:

- build an event step by step (so we don't need to pass a large amount
  of free variables around)
- integrate some default "data sources" to simplify the data analysis
  based on the events (e.g. one that adds the IOx GIT hash)
- add an abstract "emitter" concept (mostly this will send data to
  InfluxDB but this abstraction is helpful for testing of if we ever
  just want to write the events as logs)

See `event_emitter/README.md` for a short usage example.

The events are type-tagged their measurements. We'll likely have methods
with multiple event types (e.g. the query and an ingester request) and I
just want to avoid that people attach the wrong tags/fields to the wrong
events.